Rename file to apple_wallet
