+++ indico/modules/events/registration/apple_wallet.py
